// src/types/todo.tsx
export interface Todo {
      name: string;
      completed: Boolean;
}
export interface TodoState {
      todoList: Todo[];
}

// export interface TodoState {
//       todoList: Todo[];
// }

// export interface Employee {
//       string Code;
// }

// export interface Manager extends Employee {

// }

// type ButtonType = "create" | "delete"